[REFACTOR] push전 점검
