+++ src/test/java/ladder/domain/line/LadderLineTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.line;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -8,7 +7,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class BasicLadderLineTest {
+class LadderLineTest {
     @ValueSource(ints = {
             10, 100, 1000
     })
