+++ src/test/java/ladder/domain/player/PlayersTest.java
@@ -10,7 +10,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class PlayersTest {
     @ValueSource(strings = {"AA,BB"})
