+++ src/test/java/ladder/domain/point/UnconnectedRandomDirectionCreateStrategyTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.point;
 
-import ladder.core.DirectionCreateStrategy;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
